requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
tensorflow==2.15.0
scikit-learn==1.4.0
plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.0